modulize getErrorSummaryItems
